Forenkler fixtures for tester og bruker placeholder for innhold
