+++ pensjon-brevbaker/src/test/kotlin/no/nav/pensjon/etterlatte/fixtures/BarnepensjonOpphoer.kt
@@ -3,7 +3,7 @@ package no.nav.pensjon.etterlatte.fixtures
 import no.nav.pensjon.etterlatte.maler.barnepensjon.opphoer.BarnepensjonOpphoerDTO
 
 fun createBarnepensjonOpphoerDTO() = BarnepensjonOpphoerDTO(
-    innhold = emptyList(),
+    innhold = createPlaceholderForRedigerbartInnhold(),
     bosattUtland = true,
     brukerUnder18Aar = true,
 )
