--- pensjon-brevbaker/src/test/kotlin/no/nav/pensjon/etterlatte/fixtures/BarnepensjonRevurderingRedigerbartUtfallDTO.kt
@@ -1,7 +1,0 @@
-package no.nav.pensjon.etterlatte.fixtures
-
-import no.nav.pensjon.etterlatte.maler.barnepensjon.revurdering.BarnepensjonRevurderingRedigerbartUtfallDTO
-
-fun createBarnepensjonRevurderingRedigerbartUtfallDTO() = BarnepensjonRevurderingRedigerbartUtfallDTO(
-    erEtterbetaling = true,
-)
