+++ pensjon-brevbaker/src/test/kotlin/no/nav/pensjon/etterlatte/fixtures/OmstillingsstoenadOpphoer.kt
@@ -1,7 +1,15 @@
 package no.nav.pensjon.etterlatte.fixtures
 
+import no.nav.pensjon.etterlatte.maler.*
+import no.nav.pensjon.etterlatte.maler.omstillingsstoenad.opphoer.OmstillingsstoenadOpphoerDTO
 import no.nav.pensjon.etterlatte.maler.omstillingsstoenad.opphoer.OmstillingsstoenadOpphoerRedigerbartUtfallDTO
 import java.time.LocalDate
+
+fun createOmstillingsstoenadOpphoerDTO() =
+    OmstillingsstoenadOpphoerDTO(
+        innhold = createPlaceholderForRedigerbartInnhold(),
+        bosattUtland = false
+    )
 
 fun createOmstillingsstoenadpphoerRedigerbartUtfallDTO() =
     OmstillingsstoenadOpphoerRedigerbartUtfallDTO(
